requirements.txt
jieba==0.42.1
lancedb==0.30.1
langchain_core==1.2.21
langchain_openai==1.1.12
langgraph==1.1.3
marko==2.2.2
pandas==3.0.1
pydantic==2.12.5
python-dotenv==1.2.2
PyYAML==6.0.3
Requests==2.32.5
